Replace License loading by repo-method
- Use LicenseRepository::findById instead of direct class instantiation or the ModelFactory
- Replace License property-access by dedicated methods

+++ src/Module/Application/Admin/License/DeleteAction.php
@@ -26,9 +26,9 @@ declare(strict_types=0);
 namespace Ampache\Module\Application\Admin\License;
 
 use Ampache\Config\ConfigContainerInterface;
-use Ampache\Repository\Model\ModelFactoryInterface;
 use Ampache\Module\Application\ApplicationActionInterface;
 use Ampache\Module\Application\Exception\AccessDeniedException;
+use Ampache\Module\Application\Exception\ObjectNotFoundException;
 use Ampache\Module\Authorization\AccessLevelEnum;
 use Ampache\Module\Authorization\GuiGatekeeperInterface;
 use Ampache\Module\Util\UiInterface;
@@ -36,6 +36,9 @@ use Ampache\Repository\LicenseRepositoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
+/**
+ * Deletes a single license
+ */
 final class DeleteAction implements ApplicationActionInterface
 {
     public const REQUEST_KEY = 'delete';
@@ -44,19 +47,15 @@ final class DeleteAction implements ApplicationActionInterface
 
     private ConfigContainerInterface $configContainer;
 
-    private ModelFactoryInterface $modelFactory;
-
     private LicenseRepositoryInterface $licenseRepository;
 
     public function __construct(
         UiInterface $ui,
         ConfigContainerInterface $configContainer,
-        ModelFactoryInterface $modelFactory,
         LicenseRepositoryInterface $licenseRepository
     ) {
         $this->ui                = $ui;
         $this->configContainer   = $configContainer;
-        $this->modelFactory      = $modelFactory;
         $this->licenseRepository = $licenseRepository;
     }
 
@@ -66,7 +65,12 @@ final class DeleteAction implements ApplicationActionInterface
             throw new AccessDeniedException();
         }
 
-        $license = $this->modelFactory->createLicense((int)($request->getQueryParams()['license_id'] ?? 0));
+        $licenseId = (int) ($request->getQueryParams()['license_id'] ?? 0);
+
+        $license = $this->licenseRepository->findById($licenseId);
+        if ($license === null) {
+            throw new ObjectNotFoundException($licenseId);
+        }
 
         $this->licenseRepository->delete($license);
